Fix webpack.config.js
