+++ webpack.config.js
@@ -1,43 +1,36 @@
 const path = require('path');
-const { WebpackManifestPlugin } = require('webpack-manifest-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ESLintPlugin = require('eslint-webpack-plugin');
 
-const options = {};
 module.exports = {
-  entry: {
-    index: './src/index.js',
-  },
+  mode: 'development',
+  entry: './src/index.js',
+  devtool: 'inline-source-map',
   devServer: {
-    historyApiFallback: true,
-    contentBase: './dist',
+    contentBase: '/.dist',
   },
   plugins: [
     new HtmlWebpackPlugin({
+      title: 'Wbpack Exercise',
+      filename: 'index.html',
       template: './src/index.html',
     }),
-    new WebpackManifestPlugin(options),
-    new ESLintPlugin(options),
   ],
-  output: {
-    filename: '[name].js',
-    /* eslint-disable */
-    path: path.resolve(__dirname, "dist"),
-    clean: true,
-    /* eslint-disable */
-    publicPath: "/",
-  },
   module: {
     rules: [
       {
         test: /\.css$/i,
-        use: ["style-loader", "css-loader"],
+        use: ['style-loader', 'css-loader'],
       },
       {
         test: /\.(png|svg|jpg|jpeg|gif)$/i,
-        type: "asset/resource",
+        type: 'asset/resource',
       },
     ],
   },
-  mode: "development",
+  output: {
+    filename: 'bundle.js',
+    /* eslint-disable */
+    path: path.resolve(__dirname, "dist"),
+    clean: true,
+  },
 };
